Implement `FlagState` for able to manage the flag state, like validate or invalidate the flag.
