+++ osu.Game.Rulesets.Karaoke.Tests/Objects/Workings/HitObjectWorkingPropertyValidatorTest.cs
@@ -1,151 +1,8 @@
 // Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Linq;
-using NUnit.Framework;
-using osu.Game.Rulesets.Karaoke.Objects.Workings;
-
 namespace osu.Game.Rulesets.Karaoke.Tests.Objects.Workings;
 
 public class HitObjectWorkingPropertyValidatorTest
 {
-    [TestCase(default, new[] { TestEnum.Enum0, TestEnum.Enum1, TestEnum.Enum2 })]
-    [TestCase(TestEnum.Enum0, new[] { TestEnum.Enum1, TestEnum.Enum2 })]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1, new[] { TestEnum.Enum2 })]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1 | TestEnum.Enum2, new TestEnum[] { })]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum2, new[] { TestEnum.Enum1 })] // Should work without continuous values.
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum0, new[] { TestEnum.Enum1, TestEnum.Enum2 })] // Test edge case.
-    public void TestInvalidate(TestEnum invalidFlags, TestEnum[] expectedValue)
-    {
-        var validator = new HitObjectWorkingPropertyValidator<TestEnum>();
-        validator.ValidateAll();
-
-        // check the value.
-        validator.Invalidate(invalidFlags);
-        Assert.AreEqual(expectedValue, validator.GetAllValidFlags());
-
-        // value should not changed if did the same action.
-        validator.Invalidate(invalidFlags);
-        Assert.AreEqual(expectedValue, validator.GetAllValidFlags());
-
-        // should not be negative if remove the value from the validator.
-        validator.InvalidateAll();
-        validator.Invalidate(invalidFlags);
-        Assert.AreEqual(Array.Empty<TestEnum>(), validator.GetAllValidFlags());
-        Assert.AreEqual(0, validator.GetAllValidFlags().Sum(x => Convert.ToInt32(x)));
-    }
-
-    [TestCase(TestEnum.Enum0)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1 | TestEnum.Enum2)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum2)] // Should work without continuous values.
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum0)] // Test edge case.
-    public void TestInvalidateAll(TestEnum defaultFlags)
-    {
-        var validator = new HitObjectWorkingPropertyValidator<TestEnum>();
-        validator.Validate(defaultFlags);
-
-        // check the value.
-        validator.InvalidateAll();
-        Assert.AreEqual(Array.Empty<TestEnum>(), validator.GetAllValidFlags());
-        Assert.AreEqual(0, validator.GetAllValidFlags().Sum(x => Convert.ToInt32(x)));
-    }
-
-    [TestCase(default, new TestEnum[] { })]
-    [TestCase(TestEnum.Enum0, new[] { TestEnum.Enum0 })]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1, new[] { TestEnum.Enum0, TestEnum.Enum1 })]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1 | TestEnum.Enum2, new[] { TestEnum.Enum0, TestEnum.Enum1, TestEnum.Enum2 })]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum2, new[] { TestEnum.Enum0, TestEnum.Enum2 })] // Should work without continuous values.
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum0, new[] { TestEnum.Enum0 })] // Test edge case.
-    [TestCase(TestEnum.Enum2 | TestEnum.Enum2, new[] { TestEnum.Enum2 })] // Test edge case.
-    public void TestValidate(TestEnum validateFlags, TestEnum[] expectedValue)
-    {
-        var validator = new HitObjectWorkingPropertyValidator<TestEnum>();
-
-        // check the value.
-        validator.Validate(validateFlags);
-        Assert.AreEqual(expectedValue, validator.GetAllValidFlags());
-
-        // value should not changed if did the same action.
-        validator.Validate(validateFlags);
-        Assert.AreEqual(expectedValue, validator.GetAllValidFlags());
-
-        // should not exceed sum values if remove the value from the validator.
-        validator.ValidateAll();
-        validator.Validate(validateFlags);
-        Assert.AreEqual(Enum.GetValues<TestEnum>(), validator.GetAllValidFlags());
-        Assert.AreEqual(Enum.GetValues<TestEnum>().Sum(x => Convert.ToInt32(x)), validator.GetAllValidFlags().Sum(x => Convert.ToInt32(x)));
-    }
-
-    [TestCase(TestEnum.Enum0)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum1 | TestEnum.Enum2)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum2)] // Should work without continuous values.
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum0)] // Test edge case.
-    public void TestValidateAll(TestEnum validateFlags)
-    {
-        var validator = new HitObjectWorkingPropertyValidator<TestEnum>();
-
-        // check the value.
-        validator.ValidateAll();
-        Assert.AreEqual(Enum.GetValues<TestEnum>(), validator.GetAllValidFlags());
-        Assert.AreEqual(Enum.GetValues<TestEnum>().Sum(x => Convert.ToInt32(x)), validator.GetAllValidFlags().Sum(x => Convert.ToInt32(x)));
-    }
-
-    [TestCase(TestEnum.Enum0, TestEnum.Enum0, true)]
-    [TestCase(TestEnum.Enum0, TestEnum.Enum1, false)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum2, TestEnum.Enum0, true)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum2, TestEnum.Enum1, false)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum2, TestEnum.Enum2, true)]
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum0, TestEnum.Enum0, true)] // Test edge case.
-    [TestCase(TestEnum.Enum0 | TestEnum.Enum0, TestEnum.Enum2, false)] // Test edge case.
-    [TestCase(TestEnum.Enum2 | TestEnum.Enum2, TestEnum.Enum0, false)] // Test edge case.
-    [TestCase(TestEnum.Enum2 | TestEnum.Enum2, TestEnum.Enum2, true)] // Test edge case.
-    public void TestIsValid(TestEnum validateFlags, TestEnum validFlag, bool expectedValue)
-    {
-        var validator = new HitObjectWorkingPropertyValidator<TestEnum>();
-
-        // check the value.
-        validator.Validate(validateFlags);
-        Assert.AreEqual(expectedValue, validator.IsValid(validFlag));
-    }
-
-    [Test]
-    public void TestGetAllValidFlags()
-    {
-        var validator = new HitObjectWorkingPropertyValidator<TestEnum>();
-
-        // Should be possible to get all tags.
-        validator.ValidateAll();
-        Assert.AreEqual(Enum.GetValues<TestEnum>(), validator.GetAllValidFlags());
-
-        // Should not be possible to get any tags.
-        validator.InvalidateAll();
-        Assert.AreEqual(Array.Empty<TestEnum>(), validator.GetAllValidFlags());
-    }
-
-    [Test]
-    public void TestGetAllInvalidFlags()
-    {
-        var validator = new HitObjectWorkingPropertyValidator<TestEnum>();
-
-        // Should be possible to get all tags.
-        validator.ValidateAll();
-        Assert.AreEqual(Array.Empty<TestEnum>(), validator.GetAllInvalidFlags());
-
-        // Should not be possible to get any tags.
-        validator.InvalidateAll();
-        Assert.AreEqual(Enum.GetValues<TestEnum>(), validator.GetAllInvalidFlags());
-    }
-
-    [Flags]
-    public enum TestEnum
-    {
-        Enum0 = 1,
-
-        Enum1 = 1 << 1,
-
-        Enum2 = 1 << 2,
-    }
 }
